+++ tests/simpletest.py
@@ -27,7 +27,6 @@ fancyHelloWorld = (
 	"\033[0m\033[4m7\033[0m\033[46m🍄\033[0m\033[44m🎃\033[0m\033[45m🐦\033[0m"
 )
 fredHappyfaceHD = Path(f"{THISDIR}/data/fredHappyfaceHD.txt").read_text(encoding="utf-8")
-metprintOutput = Path(f"{THISDIR}/data/metprintOutput.txt").read_text(encoding="utf-8")
 cat = Path(f"{THISDIR}/data/cat.txt").read_text(encoding="utf-8")
 sys.stdout = oldStdout
 
@@ -36,33 +35,27 @@ print(fancyHelloWorld)
 print()
 print(fredHappyfaceHD)
 print()
-print(metprintOutput)
-print()
 print(cat)
 print()
 
 # To SVG
 ansiToSVG(fancyHelloWorld, f"{THISDIR}/data/fancyHelloWorld.svg", wide=False)
 ansiToSVG(fredHappyfaceHD, f"{THISDIR}/data/fredHappyfaceHD.svg")
-ansiToSVG(metprintOutput, f"{THISDIR}/data/metprintOutput.svg", wide=False)
 ansiToSVG(cat, f"{THISDIR}/data/cat.svg", width=100)
 
 # To Render
 ansiToRender(fancyHelloWorld, f"{THISDIR}/data/fancyHelloWorld.png", wide=False)
 ansiToRender(fredHappyfaceHD, f"{THISDIR}/data/fredHappyfaceHD.png")
-ansiToRender(metprintOutput, f"{THISDIR}/data/metprintOutput.png", wide=False)
 ansiToRender(cat, f"{THISDIR}/data/cat.png", width=100)
 
 # To SVGRender
 ansiToSVGRender(fancyHelloWorld, f"{THISDIR}/data/fancyHelloWorld.svg.png", wide=False)
 ansiToSVGRender(fredHappyfaceHD, f"{THISDIR}/data/fredHappyfaceHD.svg.png")
-ansiToSVGRender(metprintOutput, f"{THISDIR}/data/metprintOutput.svg.png", wide=False)
 ansiToSVGRender(cat, f"{THISDIR}/data/cat.svg.png", width=100)
 
 # To HTML
 ansiToHTML(fancyHelloWorld, f"{THISDIR}/data/fancyHelloWorld.html", wide=False)
 ansiToHTML(fredHappyfaceHD, f"{THISDIR}/data/fredHappyfaceHD.html")
-ansiToHTML(metprintOutput, f"{THISDIR}/data/metprintOutput.html", wide=False)
 ansiToHTML(cat, f"{THISDIR}/data/cat.html", width=100)
 
 # To HTMLRender
@@ -70,3 +63,8 @@ ansiToHTMLRender(fancyHelloWorld, f"{THISDIR}/data/fancyHelloWorld.html.png", wi
 ansiToHTMLRender(fredHappyfaceHD, f"{THISDIR}/data/fredHappyfaceHD.html.png")
 ansiToHTMLRender(metprintOutput, f"{THISDIR}/data/metprintOutput.html.png", wide=False)
 ansiToHTMLRender(cat, f"{THISDIR}/data/cat.html.png", width=100)
+
+# To SVGRender 06/03/2023
+ansiToSVGRender(fancyHelloWorld, f"{THISDIR}/data/hw20.svg.png", width=20)
+ansiToSVGRender(fancyHelloWorld, f"{THISDIR}/data/hwT20.svg.png", wide=True, width=20)
+ansiToSVGRender(fancyHelloWorld, f"{THISDIR}/data/hw40test.svg.png", width=40, title="test")